+++ src/server/utils.ts
@@ -1,0 +1,15 @@
+import { XTypes } from "@vex-chat/types";
+
+interface ICensoredUser {
+    userID: string;
+    username: string;
+    lastSeen: Date;
+}
+
+export const censorUser = (user: XTypes.SQL.IUser): ICensoredUser => {
+    return {
+        userID: user.userID,
+        username: user.username,
+        lastSeen: user.lastSeen,
+    };
+};
